converter/requirements.txt
astroid==2.12.13
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==4.4.2
dill==0.3.6
dnspython==2.2.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
isort==5.10.1
jedi==0.18.2
lazy-object-proxy==1.8.0
mccabe==0.7.0
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.23.5
parso==0.8.3
pathspec==0.10.2
pika==1.3.1
Pillow==9.3.0
platformdirs==2.5.4
proglog==0.1.10
pylint==2.15.6
pymongo==4.3.3
python-dotenv==0.21.0
requests==2.28.1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
